Disable model download link on job page if model inconsistent
When the model was updated after a job was submitted,
the model is no longer consistent with the one used for the job.
Thus, we now disable the download button in these cases.
Also, we no longer include model arguments in arguments on job page as
we have no way of telling whether model arguments are the ones used for
job since `upload_date` is not updated when patching arguments.