Implement OAuth flow for native clients
The ways it's currently implemented is as follows:
- The native client opens `/login` with parameters `provider` set to the name of the OAuth provider, `nc_id` with the id of the native client (currently, only `com.gams.miro` is whitelisted), `nc_redirect_uri` with the redirect URI back to the native client and `nc_public_key` with the RSA-OAEP public key of the native client.
- Engine UI starts normal OAuth flow and gets Engine access token for the provider specified but remembers the native client parameters
- After successfully retrieving Engine access token, the UI creates a AES256-GCM key and encrypts the Engine token with that. It then encrypts the AES key with the RSA-OAEP public key of the native client
- It then redirects to the URL: `${nc_id}:${nc_redirect_uri}?jwt=${encryptedJWT.data}&aes_key=${encryptedJWT.aes_key}&aes_iv=${encryptedJWT.aes_iv}` and the native client can decrypt the JWT token

This flow is not from an official RFC, but I could not find an RFC that describes how to implement OAuth for Desktop applications. The Auth0 article: https://auth0.com/blog/securing-electron-applications-with-openid-connect-and-oauth-2/ recommends opening the OAuth provider login page in a native electron window, which clearly violates the "best practices" for native clients: https://auth0.com/blog/oauth-2-best-practices-for-native-apps/ that states the login URL should be opened in the native browser.
One potential flaw in the current implementation is that any malware installed on the user's system can register the custom `com.gams.miro` URL, but with the public key implementation at least it can be ensured that whoever initiated the request is also the recipient of the auth token since noone else can decrypt the AES key and thus noone else can decrypt the Engine auth token. So at least there is no man-in-the-middle attack possibility between user's browser and native client.